+++ Assets/OxGKit/ActionSystem/Scripts/Runtime/Logger.cs
@@ -1,0 +1,7 @@
+﻿using OxGKit.LoggingSystem;
+
+namespace OxGKit.ActionSystem
+{
+    [LoggerName("OxGKit.ActionSystem.Logger")]
+    public class Logger : Logging { }
+}