+++ Assets/OxGKit/InputSystem/Scripts/Runtime/Logger.cs
@@ -1,0 +1,7 @@
+﻿using OxGKit.LoggingSystem;
+
+namespace OxGKit.InputSystem
+{
+    [LoggerName("OxGkit.InputSystem.Logger")]
+    public class Logger : Logging { }
+}