+++ Assets/OxGKit/NoticeSystem/Scripts/Runtime/Logger.cs
@@ -1,0 +1,7 @@
+﻿using OxGKit.LoggingSystem;
+
+namespace OxGKit.NoticeSystem
+{
+    [LoggerName("OxGKit.NoticeSystem.Logger")]
+    public class Logger : Logging { }
+}