+++ Assets/OxGKit/NoticeSystem/Scripts/Runtime/NoticeItem.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using OxGKit.LoggingSystem;
+using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 
@@ -139,7 +140,7 @@ namespace OxGKit.NoticeSystem
                     if (!this.gameObject.activeSelf)
                     {
                         this.gameObject.SetActive(true);
-                        Debug.Log($"<color=#6dedff>[{nameof(NoticeSystem)}] <color=#94ff2c>(Match Condition)</color> Notify Notice Item <color=#94ff2c>{this.name}</color></color>");
+                        Logging.Print<Logger>($"<color=#6dedff>[{nameof(NoticeSystem)}] <color=#94ff2c>(Match Condition)</color> Notify Notice Item <color=#94ff2c>{this.name}</color></color>");
                     }
                     return;
                 }