+++ Assets/OxGKit/ActionSystem/Scripts/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## CHANGELOG
 
-## [0.0.2-preview] - 2023-05-21=8
+## [0.0.3-preview] - 2023-08-27
+- Added Logger by LoggingSystem.
+
+## [0.0.2-preview] - 2023-05-21
 - Fixed OnDone invoke timing (If call MarkAsDone will invoke OnDone).
 - Modified done flags are IsDone and IsAllDone.