+++ Assets/OxGKit/InputSystem/Scripts/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## CHANGELOG
 
+## [0.0.2-preview] - 2023-08-27
+- Added Logger by LoggingSystem.
+
 ## [0.0.1-preview] - 2023-05-21
 - Added InputSystem.